pyproject.toml
[project]
name = "attendance-system-fastapi"
version = "0.0.1"
description = ""
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi==0.136.0",
    "uvicorn==0.45.0",
    "qrcode==8.2",
    "pillow==12.2.0",
    "openpyxl==3.1.5",
    "pydantic==2.13.3",
    "pydantic-core==2.46.3",
    "python-dotenv==1.2.2",
    "pyinstaller==6.19.0",
    "httpx==0.28.1",
    "pytest==9.0.3",
    "pytest-asyncio==1.3.0",
    "gspread==6.2.1",
    "google-auth==2.49.2",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
